python-app/requirements.txt
Flask==3.1.0
numpy==1.26.4
pandas==2.2.3
langchain==0.3.16
langchain-community==0.3.16
PyYAML==6.0.2
requests
regex==2024.11.6
ultralytics==8.3.69
boto3==1.36.11
botocore==1.36.11
openai==1.60.2
prometheus-flask-exporter